completed the tax service

+++ api/internal/dto/tax/tax.go
@@ -2,5 +2,5 @@ package tax
 
 type Tax struct {
 	State  string
-	Amount float32
+	Amount float64
 }
